feat: add pixel font

+++ _build/Debug.h
@@ -7,7 +7,7 @@ class Debug
 private:
 	Game* gameInstance = nullptr;
 	Vector2 position;
-	int fontSize = 24;
+	float fontSize = 24.0f;
 	void drawMapCollisions();
 	void drawPlayerState();
 	void drawPlayerInfo();
